--- myCookingManager/backend/html_email_list.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html>
-<body>
-
-    <ul>
-        origin
-        <!-- <li style="list-style-type: none;"><input id="ingredient" type="checkbox" style="width: 12px; height: 12px; position: relative; vertical-align: bottom; margin-right: 10px;"/><label for="ingredient" style="font-size: 16px;">${ingredient}</label></li> -->
-    </ul>
-
-
-</body>
-
-</html>
