+++ myCookingManager/backend/utilities/constants.js
