+++ myCookingManager/backend/utilities/html_email_list.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html>
+<body>
+
+    <ul>
+        origin
+        <!-- <li style="list-style-type: none;"><input id="ingredient" type="checkbox" style="width: 12px; height: 12px; position: relative; margin-right: 10px; top: 1px;"/><label for="ingredient" style="font-size: 16px;">${ingredient}</label></li> -->
+    </ul>
+
+
+</body>
+
+</html>
